solution formatting
