+++ client/components/index.js
@@ -17,4 +17,4 @@ export {default as Topics} from './Topics'
 export {default as DashboardPrompt} from './DashboardPrompt'
 export {default as DashboardStats} from './DashboardStats'
 export {default as Examples} from './Examples'
-export {default as Solution} from './Solution'
+export {default as Solutions} from './Solutions'
